Update CommentBusiness and CommentsController
