+++ OngProject/appsettings.json
@@ -8,6 +8,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DefaultConnection": "Server=JOHAN;Database=ONGProject;Trusted_Connection=True;"
+    "DefaultConnection": ""
   }
 }
